refactor(playground): extract strategy-picker, scenario-picker, tweak-drawer
Three largest feature extractions from main.ts (633 lines removed):

features/strategy-picker/ — dispatch + reposition popovers, labels,
  renderPopoverOptions, attachOutsideClickForPopovers
features/scenario-picker/ — scenario cards, switch logic, mobile
  sheet-compact readout, reconcileStrategyWithScenario
features/tweak-drawer/ — parameter rendering, stepper logic,
  hot-swap, TweakRowHandles

Each feature defines narrow local interface types and accepts
resetAll as a callback to avoid circular imports.
