+++ playground/src/features/strategy-picker/index.ts
@@ -1,0 +1,15 @@
+export { STRATEGY_LABELS } from "./labels";
+export {
+  renderPaneStrategyInfo,
+  refreshStrategyPopovers,
+  isAnyStrategyPopoverOpen,
+  attachStrategyPopover,
+} from "./dispatch-popover";
+export {
+  renderPaneRepositionInfo,
+  refreshRepositionPopovers,
+  isAnyRepositionPopoverOpen,
+  closeAllPopovers,
+  attachRepositionPopover,
+  attachOutsideClickForPopovers,
+} from "./reposition-popover";
