+++ playground/src/features/tweak-drawer/handles.ts
@@ -1,0 +1,11 @@
+export interface TweakRowHandles {
+  root: HTMLElement;
+  value: HTMLElement;
+  defaultV: HTMLElement;
+  dec: HTMLButtonElement;
+  inc: HTMLButtonElement;
+  reset: HTMLButtonElement;
+  trackFill: HTMLElement | null;
+  trackDefault: HTMLElement | null;
+  trackThumb: HTMLElement | null;
+}
